requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dlib-bin==19.24.6
face_recognition @ git+https://github.com/thetoby9944/face_recognition.git
face_recognition_models==0.3.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
ImageHash==4.3.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==2.0.1
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
soupsieve==2.7
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
warcio==1.7.5
watchdog==6.0.0
